adbook-backend/requirements.txt
# Django Core
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
django-extensions==3.2.3
django-storages==1.14.2

# Database & Cache
psycopg==3.2.4
redis==5.0.1


# ASGI & WebSockets
channels==4.1.0
channels-redis==4.2.0

# Celery
celery[redis]==5.4.0

# Auth & Security
django-allauth==0.63.2
pyotp==2.9.0
qrcode[pil]==7.4.2

# Utils
Pillow==10.4.0
python-decouple==3.8
gunicorn[gevent]==22.0.0
uvicorn[standard]==0.30.3

# Testing
pytest==8.2.1
pytest-django==4.8.0
factory-boy
pytest-asyncio==0.23.8

# Linting & Formatting
black==24.8.0
flake8==7.1.0
isort==5.13.2
mypy==1.11.1

# Production
sentry-sdk==2.13.0
newrelic
